// frontend/src/api/generated/model/postDetailResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * 先進プロジェクト実験 Team2 避難所 API
 * OpenAPI spec version: 1.0.0
 */
import type { PostDetailResponseStatus } from './postDetailResponseStatus';
import type { MediaItem } from './mediaItem';
import type { LocationTrackPoint } from './locationTrackPoint';

/**
 * 投稿詳細情報
 */
export interface PostDetailResponse {
  /** 投稿ID */
  id: string;
  /** 避難所ID */
  shelterId: number;
  /** 避難所名 */
  shelterName?: string;
  /** 投稿者の表示名 */
  authorName: string;
  /**
   * 投稿本文
   * @nullable
   */
  content?: string | null;
  /**
   * 災害等の事象が発生した時刻
   * @nullable
   */
  occurredAt?: string | null;
  /** 投稿された時刻 */
  postedAt: string;
  /**
   * 現在の状況ステータス
   * @nullable
   */
  status?: PostDetailResponseStatus;
  /** 添付されたメディア情報 */
  media: MediaItem[];
  /** 位置情報の時系列データ */
  locationTrack?: LocationTrackPoint[];
  /** コメント数 */
  commentCount: number;
}